config: Cast price_stop_level to (float)

+++ config/EURUSD_M1.h
@@ -10,7 +10,7 @@ struct Stg_TMA_Band_SVE_True_EURUSD_M1_Params : Stg_TMA_Band_SVE_True_Params {
     TMA_Band_SVE_True_SignalCloseMethod = signal_close_method = 0;
     TMA_Band_SVE_True_SignalCloseLevel = signal_close_level = (float)0;
     TMA_Band_SVE_True_PriceStopMethod = price_stop_method = 0;
-    TMA_Band_SVE_True_PriceStopLevel = price_stop_level = 2;
+    TMA_Band_SVE_True_PriceStopLevel = price_stop_level = (float)2;
     TMA_Band_SVE_True_TickFilterMethod = tick_filter_method = 1;
     TMA_Band_SVE_True_MaxSpread = max_spread = 0;
   }
